refactor: 사용안하는 파일 제거
