--- src/components/Buttons/index.jsx
@@ -1,3 +1,0 @@
-import { Button } from './style';
-
-export default Button;
