--- src/components/Message/CoinMessage/style.js
@@ -1,17 +1,0 @@
-import styled from '@emotion/styled';
-
-const CurrentCoin = styled.span`
-  transition: all 0.3s ease-in-out;
-  color: ${(props) => (props.isVisible ? 'var(--color-title-text)' : 'var(--color-yellow)')};
-`;
-
-const EarnedCoin = styled.span`
-  transition: all 0.3s ease-in-out;
-  margin-left: ${(props) => (props.isVisible ? '4px' : 0)};
-  color: var(--color-yellow);
-  transform: ${(props) => (props.isVisible ? 'translateX(0)' : 'translateX(-100%)')};
-  opacity: ${(props) => (props.isVisible ? 1 : 0)};
-  width: ${(props) => (props.isVisible ? `${props.width}px` : 0)};
-`;
-
-export default { CurrentCoin, EarnedCoin };
