--- src/components/SignForm/SignUp/SignUpFirst.jsx
@@ -1,188 +1,0 @@
-import { useRef, useState, useEffect, useReducer } from 'react';
-import S from '../style';
-import LogoWhite from '@/assets/images/logo.season2.white.svg';
-import Button from '@/components/Buttons';
-import { BlackOverlay } from '@/components/BlackOverlay';
-import AlertModal from '@/components/Modal/AlertModal';
-import {
-  textReducer,
-  focusReducer,
-  isValidEmail,
-  checkFormValidity,
-  setInputState,
-} from '@/utils/SignUp';
-import { SIGNUP_PLACEHOLDER } from '@/constants/SignUp';
-
-const SignUpFirst = ({ setFormType, onNext }) => {
-  const [inputState, dispatchText] = useReducer(textReducer, {
-    githubId: '',
-    password: '',
-    checkedPassword: '',
-    name: '',
-    studentId: '',
-    baekjoonId: '',
-    discordId: '',
-    notionEmail: '',
-  });
-  const [focusState, dispatchFocus] = useReducer(focusReducer, {
-    githubIdFocus: false,
-    passwordFocus: false,
-    checkedPasswordFocus: false,
-    nameFocus: false,
-    studentIdFocus: false,
-    beakjoonIdFocus: false,
-    discordIdFocus: false,
-    notionEmailFocus: false,
-  });
-  const [isOpenedModal, setIsOpenedModal] = useState(false);
-  const [modalText, setModalText] = useState('');
-
-  const githubIdRef = useRef(null);
-  const passwordRef = useRef(null);
-  const checkedPasswordRef = useRef(null);
-  const nameRef = useRef(null);
-  const studentIdRef = useRef(null);
-  const baekjoonIdRef = useRef(null);
-  const discordIdRef = useRef(null);
-  const notionEmailRef = useRef(null);
-
-  const inputObject = {
-    GITHUB_ID: {
-      inputState: inputState.githubId,
-      focusState: focusState.githubIdFocus,
-      ref: githubIdRef,
-      check: () => inputState.githubId.length <= 1,
-    },
-    PASSWORD: {
-      type: 'password',
-      inputState: inputState.password,
-      focusState: focusState.passwordFocus,
-      ref: passwordRef,
-      check: () => inputState.password.length <= 3,
-    },
-    CHECKED_PASSWORD: {
-      type: 'password',
-      inputState: inputState.checkedPassword,
-      focusState: focusState.checkedPasswordFocus,
-      ref: checkedPasswordRef,
-      check: () => !(inputState.password === inputState.checkedPassword),
-    },
-    NAME: {
-      inputState: inputState.name,
-      focusState: focusState.nameFocus,
-      ref: nameRef,
-      check: () => inputState.name.length <= 1,
-    },
-    STUDENT_ID: {
-      inputState: inputState.studentId,
-      focusState: focusState.studentIdFocus,
-      ref: studentIdRef,
-      check: () => inputState.studentId.length !== 10,
-    },
-    BAEKJOON_ID: {
-      inputState: inputState.baekjoonId,
-      focusState: focusState.baekjoonIdFocus,
-      ref: baekjoonIdRef,
-      check: () => inputState.baekjoonId.length <= 1,
-    },
-    DISCORD_ID: {
-      inputState: inputState.discordId,
-      focusState: focusState.discordIdFocus,
-      ref: discordIdRef,
-      check: () => inputState.discordId.length <= 1,
-    },
-    NOTION_EMAIL: {
-      inputState: inputState.notionEmail,
-      focusState: focusState.notionEmailFocus,
-      ref: notionEmailRef,
-      check: () => !isValidEmail(inputState.notionEmail),
-    },
-  };
-
-  useEffect(() => {
-    Object.entries(inputObject).map(([key, value]) => {
-      if (value.ref) {
-        value.ref.current.placeholder = SIGNUP_PLACEHOLDER[key].DEFAULT_PLACEHOLDER;
-      }
-    });
-  }, []);
-
-  const onChange = (e) => {
-    const { name, value } = e.target;
-    dispatchText({ type: name, text: value });
-  };
-
-  const onSubmit = async (event) => {
-    event.preventDefault();
-    if (checkFormValidity(inputObject, dispatchText, dispatchFocus)) {
-      setInputState(inputState);
-      onNext();
-    }
-  };
-
-  const onKeyDown = (event) => {
-    if (event.key === 'Enter') {
-      event.preventDefault();
-      if (checkFormValidity(inputObject, dispatchText, dispatchFocus)) {
-        setInputState(inputState);
-        onNext();
-      }
-    }
-  };
-  return (
-    <S.SignBox onSubmit={onSubmit}>
-      <BlackOverlay isOpen={isOpenedModal} />
-      <AlertModal
-        isOpen={isOpenedModal}
-        description={modalText}
-        closeModal={() => {
-          setIsOpenedModal(false);
-          setFormType('SIGNIN');
-        }}
-      />
-      <S.ImageWrapper>
-        <S.LogoImage src={LogoWhite} />
-      </S.ImageWrapper>
-      <S.PhraseWrapper>
-        <S.Phrase>알록에 온 것을 환영해요!</S.Phrase>
-      </S.PhraseWrapper>
-      <S.BreakLine />
-      {Object.entries(inputObject).map(([key, value]) => {
-        return (
-          <S.InputBox
-            name={key}
-            ref={value.ref}
-            isFocused={value.focusState}
-            value={value.inputState}
-            onChange={onChange}
-            type={value.type === 'password' ? 'password' : 'text'}
-            onKeyDown={onKeyDown}
-          />
-        );
-      })}
-
-      <Button
-        color={'blue'}
-        buttonType={'active'}
-        size={'medium'}
-        style={{
-          height: 40,
-          textAlign: 'start',
-          padding: '4px 8px',
-          fontSize: 14,
-          borderRadius: 8,
-        }}
-      >
-        다음으로
-      </Button>
-      <S.UnderlinedText
-        onClick={() => {
-          setFormType('SIGNIN');
-        }}
-      >
-        로그인하기
-      </S.UnderlinedText>
-    </S.SignBox>
-  );
-};
-export default SignUpFirst;
