--- src/components/SignForm/SignUp/index.jsx
@@ -1,18 +1,0 @@
-import { useFunnel } from '@/hooks/useFunnel';
-import SignUpFirst from './SignUpFirst';
-import SignUpSecond from './SignUpSecond';
-
-const SignUp = ({ setFormType }) => {
-  const { Step, Funnel, setStep } = useFunnel('회원가입');
-  return (
-    <Funnel>
-      <Step name="회원가입">
-        <SignUpFirst setFormType={setFormType} onNext={() => setStep('코스선택')} />
-      </Step>
-      <Step name="코스선택">
-        <SignUpSecond setFormType={setFormType} />
-      </Step>
-    </Funnel>
-  );
-};
-export default SignUp;
